Remove PIMPL constraints, add RenderContext.
PIMPL constraints are not able to work because struct private is not
defined at the point when the destructor is called in the constraint
class.

Add render context. Driver started to become to complex, therefore
RenderContext is introduced to hold state of the current rendering
process.
The goal that driver keep in the lowest level as it is possible.

From now window's has render context and not directly the driver.
Driver should be inaccessable from outside of the rendering layer.