Library implementations started with shader library.
In order to be able to run in parallel way, first libraries must be
defined to restrict access to different managers.

To be able to remove the window and render context references must be
create the libraries: GeometryLibrary, ShaderLibrary, MaterialLibrary.

It was the first step of it a thread safe ShaderLibrary.